requirements.txt
playwright==1.40.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
